refactor(solana-contracts): unify program_id WH core bridge & token bridge, more cleanup
